# perf(parse): cache length, return early