@blakeembrey 2024-10-02
Tried rebasing on the work I was doing in https://github.com/jshttp/cookie/pull/170. The alternative in https://github.com/jshttp/cookie/pull/170 was to do a `while (index < length - 2)` check which would eliminate lines 4 & 5 from your list, as well as merges the new `if` with the `while` making it only 2/5. So at this point it's only eliminating the `dec` line, but I'll merge anyway, I'd love to see more perf wins!

Thanks for the PR.